requirements.txt
# Core ML Dependencies
transformers==4.35.0
datasets==2.14.7
torch==2.0.1
accelerate==0.20.3
peft==0.3.0
sentence-transformers==2.2.2

# ML Utilities
scikit-learn==1.2.2
rouge-score==0.1.2
sacrebleu==2.3.1

# Data Processing
pandas==2.0.3
numpy==1.24.3
tqdm==4.65.0
pyyaml==6.0

# Web Framework
streamlit==1.28.1

# API Clients
openai==1.6.1
requests==2.31.0
httpx==0.25.2

# Development
pytest==7.4.0
playwright==1.40.0